OPS-55
Sizeifyb issue with images that have space in file name - python client.

when given a URL using spaces and not %20 character sequences, the client fails to automatically convert to %20 before again urlencoding the string. It is arguable whether this is actually a bug, since spaces are not valid in URLs. But i'm going to fix it anyway.